--- a/vdr/integration_test.go
+++ b/vdr/integration_test.go
@@ -22,6 +22,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/nuts-foundation/go-stoabs"
 
 	"net/url"
 	"sync"
@@ -282,7 +283,14 @@ func setup(t *testing.T) testContext {
 	cryptoInstance.Configure(nutsConfig)
 
 	// Storage
-	storageProvider := storage.NewTestStorageEngine(testDir)
+	var memstore stoabs.KVStore
+	if memstore, err = storage.CreateTestBBoltStore(testDir + "/test.db"); err != nil {
+		t.Fatal(err)
+	}
+	storageProvider := storage.StaticKVStoreProvider{memstore}
+	t.Cleanup(func() {
+		memstore.Close(context.Background())
+	})
 
 	// DID Store
 	didStore := store.NewMemoryStore()
@@ -312,7 +320,7 @@ func setup(t *testing.T) testContext {
 		docResolver,
 		docFinder,
 		eventPublisher,
-		storageProvider.GetProvider(network.ModuleName),
+		&storageProvider,
 	)
 	nutsNetwork.Configure(nutsConfig)
 	nutsNetwork.Start()