closes #1219 

network/vdr integration test use bbolt store without syncing

test speedup:
network 26s -> 9s
vdr 5s -> 0.7s